// src/models/packageModel.ts
export default class PackageModel {
    active: boolean;
    cpu_cap: number;
    created_at: string;
    default: boolean;
    description: string;
    group: string;
    max_lwps: number;
    max_physical_memory: number;
    max_swap: number;
    name: string;
    quota: number;
    updated_at: string;
    uuid: string;
    v: number;
    version: string;
    zfs_io_priority: number;
}